make Scanner::scan_tokens no longer mutate self
Scanner::scan_tokens now instead takes a source and acts on it making
it possible to use the same Scanner to scan more than one source, which
kind of makes sense.

This was implemented by adding a new ScannerContext that encapsulates
the state required for one scan, and having the Scanner create a new
ScannerContext for each call to scan_tokens.

This commit also moves the ownership of the scanned tokens to the
caller, disconnecting the lifetime of the scanned tokens from the
Scanner.